requirements.txt
google-genai>=1.0.0
requests>=2.31.0
langgraph>=0.1.0
